Splitting the classifier into a classifier and a stitcher
