--- example-config.yml
@@ -1,1 +1,0 @@
-modeling/config/classifier-no-positional.yml
