+++ modeling/classify.py
@@ -25,41 +25,45 @@ import torch
 import yaml
 from PIL import Image
 
-from modeling import train, data_loader, negative_label
+from modeling import train, data_loader, negative_label, stitch
+
+
+DEBUG = False
 
 
 class Classifier:
 
     def __init__(self, **config):
         self.config = config
-        model_config = yaml.safe_load(open(config["model_config_file"]))
+        self.model_config = yaml.safe_load(open(config["model_config_file"]))
         # the "labels" list from the config file does not include the "negative"
         # label, but we need it here so we add it
-        self.labels = model_config['labels'] + [negative_label]
+        # TODO: do not use labels
+        self.labels = self.model_config['labels'] + [negative_label]
         self.featurizer = data_loader.FeatureExtractor(
-            img_enc_name=model_config["img_enc_name"],
-            pos_enc_name=model_config.get("pos_enc_name", None),
-            pos_enc_dim=model_config.get("pos_enc_dim", 0),
-            max_input_length=model_config.get("max_input_length", 0),
-            pos_unit=model_config.get("pos_unit", 0),
+            img_enc_name=self.model_config["img_enc_name"],
+            pos_enc_name=self.model_config.get("pos_enc_name", None),
+            pos_enc_dim=self.model_config.get("pos_enc_dim", 0),
+            max_input_length=self.model_config.get("max_input_length", 0),
+            pos_unit=self.model_config.get("pos_unit", 0),
         )
         self.classifier = train.get_net(
             in_dim=self.featurizer.feature_vector_dim(),
-            n_labels=len(model_config['bins']['pre'].keys()) + 1,
-            num_layers=model_config["num_layers"],
-            dropout=model_config["dropouts"],
+            n_labels=len(self.model_config['bins']['pre'].keys()) + 1,
+            num_layers=self.model_config["num_layers"],
+            dropout=self.model_config["dropouts"],
         )
         self.classifier.load_state_dict(torch.load(config["model_file"]))
         # TODO (krim @ 12/14/23): deal with post bin
         # self.postbin = config.get("postbin", None)
-
-        # configuration settings from the config file
         self.sample_rate = self.config["sample_rate"]
-        self.minimum_frame_score = self.config["minimum_frame_score"]
-        self.minimum_timeframe_score = self.config["minimum_timeframe_score"]
-        self.minimum_frame_count = self.config["minimum_frame_count"]
-        # debugging
-        self.dribble = False
+        self.debug = DEBUG
+
+    def __str__(self):
+        return (f"<Classifier "
+                + f'img_enc_name="{self.model_config["img_enc_name"]}" '
+                + f'pos_enc_name="{self.model_config["pos_enc_name"]}" '
+                + f'sample_rate={self.sample_rate}>')
 
     def set_parameters(self, parameters):
         """Take the parameters from the configuration file and update them with
@@ -84,7 +88,7 @@ class Classifier:
         """Loops over the frames in a video and for each frame extracts the features
         and applies the classifier. Returns a list of predictions, where each prediction
         is an instance of numpy.ndarray."""
-        if self.dribble:
+        if self.debug:
             print(f'Processing {mp4_file}...')
             print(f'Labels: {self.labels}')
         logging.info(f'processing {mp4_file}...')
@@ -102,72 +106,10 @@ class Classifier:
             features = self.featurizer.get_full_feature_vectors(img, ms, dur)
             prediction = self.classifier(features).detach()
             prediction = Prediction(ms, self.labels, prediction)
-            if self.dribble:
+            if self.debug:
                 print(prediction)
             predictions.append(prediction)
         return predictions
-
-    def extract_timeframes(self, predictions):
-        timeframes = self.collect_timeframes(predictions)
-        if self.dribble:
-            print_timeframes("Potential timeframes", timeframes)
-        timeframes = self.filter_timeframes(timeframes)
-        timeframes = self.remove_overlapping_timeframes(timeframes)
-        if self.dribble:
-            print_timeframes("Selected timeframes", timeframes)
-        return timeframes
-
-    def collect_timeframes(self, predictions: list) -> dict:
-        """Find sequences of frames for all labels where the score of each frame
-        is at least the mininum value as defined in self.minimum_frame_score."""
-        timeframes = []
-        open_frames = { label: TimeFrame(label) for label in self.labels}
-        for prediction in predictions:
-            for i, label in enumerate(prediction.labels):
-                if label == negative_label:
-                    continue
-                score = prediction.data[i]
-                if score < self.minimum_frame_score:
-                    if open_frames[label]:
-                        timeframes.append(open_frames[label])
-                    open_frames[label] = TimeFrame(label)
-                else:
-                    open_frames[label].add_point(prediction.timepoint, score)
-        for label in self.labels:
-            if open_frames[label]:
-                timeframes.append(open_frames[label])
-        for tf in timeframes:
-            tf.finish()
-        return timeframes
-
-    def filter_timeframes(self, timeframes: list) -> list:
-        """Filter out all timeframes with an average score below the threshold defined
-        in the configuration settings."""
-        # TODO: this now also uses the minimum number of samples, but maybe do this
-        # filtering later in case we want to use short competing timeframes as a way
-        # to determine whther another frame is viable
-        return [tf for tf in timeframes
-                if (tf.score > self.minimum_timeframe_score
-                    and len(tf) >= self.minimum_frame_count)]
-
-    def remove_overlapping_timeframes(self, timeframes: list) -> list:
-        all_frames = list(sorted(timeframes, key=lambda tf: tf.score, reverse=True))
-        outlawed_timepoints = set()
-        final_frames = []
-        for frame in all_frames:
-            if self.is_included(frame, outlawed_timepoints):
-                continue
-            final_frames.append(frame)
-            for p in range(frame.start, frame.end + self.sample_rate, self.sample_rate):
-                outlawed_timepoints.add(p)
-        return final_frames
-
-    def is_included(self, frame, outlawed_timepoints: set):
-        #start, end, _, _ = frame
-        for i in range(frame.start, frame.end + self.sample_rate, self.sample_rate):
-            if i in outlawed_timepoints:
-                return True
-        return False
 
     def pp(self):
         # debugging method
@@ -175,10 +117,6 @@ class Classifier:
         print(f"   sample_rate         = {self.sample_rate}")
         print(f"   minimum_frame_score = {self.minimum_timeframe_score}")
         print(f"   minimum_frame_count = {self.minimum_frame_count}")
-
-
-def softmax(x):
-    return(np.exp(x - np.max(x)) / np.exp(x - np.max(x)).sum())
 
 
 def save_predictions(predictions: list, filename: str):
@@ -210,56 +148,13 @@ def print_predictions(predictions, filename=None):
     fh.write(f'\nTOTAL PREDICTIONS: {len(predictions)}\n')
 
 
-def print_timeframes(header, timeframes: list):
-    print(f'\n{header} ({len(timeframes)})')
-    for tf in sorted(timeframes, key=lambda tf: tf.start):
-        print(tf)
-
-
-class TimeFrame:
-
-    def __init__(self, label: str):
-        self.label = label
-        self.points = []
-        self.scores = []
-        self.start = None
-        self.end = None
-        self.score = None
-
-    def __len__(self):
-        return len(self.points)
-
-    def __nonzero__(self):
-        return len(self) != 0
-
-    def __str__(self):
-        if self.is_empty():
-            return "<TimePoint empty>"
-        else:
-            return f"<TimeFrame {self.label} {self.points[0]}:{self.points[-1]} score={self.score:0.4f}>"
-
-    def add_point(self, point, score):
-        self.points.append(point)
-        self.scores.append(score)
-
-    def finish(self):
-        """Once all points have been added to a timeframe, use this method to
-        calculate the timeframe score from the points and to set start and end."""
-        self.score = sum(self.scores) / len(self)
-        self.start = self.points[0]
-        self.end = self.points[-1]
-
-    def is_empty(self):
-        return len(self) == 0
-
-
 class Prediction:
 
-    """Class to store a prediction from the model. It is meant to simplify the rest
-    of the code a bit and manage some of the intricacies of the data structures that
-    are involved. One thing it does is to run softmax over the scores in the tensor.
+    """Class to store a prediction from the model. Runs softmax over the scores in the
+    tensor and can retrieve scores for all labels.
 
     timepoint  -  the location of the frame, in milliseconds
+    labels     -  labels for the prediction, taken from the model
     tensor     -  the tensor that results from running the model on the features
     data       -  the tensor simplified into a simple list with scores for each label
 
@@ -271,8 +166,9 @@ class Prediction:
         self.labels = labels
         self.tensor = prediction
         if data is None:
-            # TODO: probably use torch.nn.Softmax()
-            self.data = softmax(self.tensor.detach().numpy()).tolist()
+            # TODO: it seems like the dimension has to be -1 or 0. Minor changes
+            # in the updated scores result, not sure which one is better. 
+            self.data = torch.nn.Softmax(dim=0)(self.tensor).detach().numpy().tolist()
         else:
             self.data = data
 
@@ -293,15 +189,20 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     conf_help = "the YAML config file"
     pred_help = "use cached predictions"
-    parser.add_argument("-i", "--input", help="input video file")
-    parser.add_argument("-c", "--config", default='modeling/config/classifier.yml', help=conf_help)
+    parser.add_argument("--input", help="input video file")
+    parser.add_argument("--config", default='modeling/config/classifier.yml', help=conf_help)
     parser.add_argument("--use-predictions", action='store_true', help=pred_help)
     parser.add_argument("--debug", action='store_true', help="turn on debugging")
     args = parser.parse_args()
 
-    classifier = Classifier(**yaml.safe_load(open(args.config)))
+    configs = yaml.safe_load(open(args.config))
+    classifier = Classifier(**configs)
+    stitcher = stitch.Stitcher(**configs)
     if args.debug:
-        classifier.dribble = True
+        classifier.debug = True
+        stitcher.debug = True
+    print(classifier)
+    print(stitcher)
 
     input_basename, extension = os.path.splitext(args.input)
     predictions_file = f'{input_basename}.json'
@@ -309,15 +210,4 @@ if __name__ == '__main__':
         predictions = load_predictions(predictions_file, classifier.labels)
     else:
         predictions = classifier.process_video(args.input)
-        #save_predictions(predictions, predictions_file)
-    #print_predictions(predictions, filename='predictions.txt')
-
-    timeframes = classifier.collect_timeframes(predictions)
-    print_timeframes('Collected frames', timeframes)
-    
-    timeframes = classifier.filter_timeframes(timeframes)
-    print_timeframes('Filtered frames', timeframes)
-
-    timeframes = classifier.remove_overlapping_timeframes(timeframes)
-    print_timeframes('Final frames', timeframes)
-
+    timeframes = stitcher.create_timeframes(predictions)
